edit and delete pop up on support staff payments

+++ finance/views.py
@@ -5,15 +5,8 @@ from django.shortcuts import render, HttpResponse,redirect
 from .models import *
 from django.db.models import Sum
 from frontend.models import *
-# from .models import Supportstaff , Staffpayments , ExpenseRecord , Fees , Bankdetails , Receipts , Supportstaffpayment , Teacherspayment , Teachers
-
-
-# Create your views here.
-#  Display_Fees = ['paymentid', 'stdnumber', 'stdname', 'studentclass', 'amount', 'balance', 'modeofpayment', 'date']
-#Display_ExpenseRecords =['expenseid', 'category', 'amountrequired', 'expensedate', 'amountpaid', 'balance']
-#displaystaffpayments = ['staffname' , 'datepaid' , 'salary' , 'amountpaid' , 'balance' , 'position' , 'bankaccnum']
-#'staffname' , 'bankname' , 'accnum' , 'accname'
-#'receiptnum' , 'transactiondate' , 'amountpaid' , 'item' , 'balance' , 'payername'
+from django.core import serializers
+from django.http import JsonResponse
 
 def editteacherpayments(request , id):
     teacherpayment = Teacherspayment.objects.get(id = id)
@@ -25,7 +18,7 @@ def editteacherpayments(request , id):
         bankaccnum = request.POST.get('bankaccnum')
         teacherid = request.POST.get('teacherid')
         teachername = request.POST.get('teachername')
-        
+
         teacherpayment.salary = salary 
         teacherpayment.amountpaid = amountpaid 
         teacherpayment.balance = balance 
@@ -33,17 +26,18 @@ def editteacherpayments(request , id):
         teacherpayment.bankaccnum = bankaccnum 
         teacherpayment.teacherid = teacherid 
         teacherpayment.teachername = teachername 
-        
+
         teacherpayment.save()
         messages.success(request , "Teacher Payment Edit Successfull")
-        
+
         return redirect('teacherpaymentslist')
-        
+
     return render(request , 'finance/staffpayments/editteacherpayments.html' , {'teacherpayment' : teacherpayment})
 
-def editsupportstaffpayment(request , paymentid):
-    staffpayment = Supportstaffpayment.objects.get(paymentid = paymentid)
+def editsupportstaffpayment(request):
     if request.method == 'POST':
+        paymentid = request.POST.get('paymentid')
+        staffpayment = Supportstaffpayment.objects.get(paymentid = paymentid)
         supportstaffid = request.POST.get('supportstaffid')
         salary = request.POST.get('salary')
         amountpaid = request.POST.get('amountpaid')
@@ -52,7 +46,7 @@ def editsupportstaffpayment(request , paymentid):
         #bankaccnum = request.POST.get('bankaccnum')
         paymentdate = request.POST.get('paymentdate')
         staffname = request.POST.get('staffname')
-        
+
         staffpayment.supportstaffid = supportstaffid
         staffpayment.salary = salary
         staffpayment.amountpaid = amountpaid 
@@ -61,9 +55,9 @@ def editsupportstaffpayment(request , paymentid):
         #staffpayment.bankaccnum = bankaccnum
         staffpayment.paymentdate = paymentdate 
         staffpayment.staffname = staffname 
-        
+
         staffpayment.save()
-        
+
         messages.success(request , 'Edited Successfully')
         return redirect('SupportstaffpaymentsLists')
     return render(request , 'finance/staffpayments/editsupportstaffpayments.html' , {'staffpayment': staffpayment})
@@ -73,15 +67,16 @@ def deleteteacherpayment(request , id):
     teacherpayment.delete()
     messages.success(request , 'Payment Deleted Successfully')
     return redirect('teacherpaymentslist')
-    
-def deletesupportstaffpayment(request , paymentid):
-    payid = Supportstaffpayment.objects.filter(paymentid = paymentid)
-    payid.delete()
-    messages.success(request , "Payment deleted successfully")
-    supportstaffinfo = Supportstaffpayment.objects.all()
-    return redirect('SupportstaffpaymentsLists')
-    
-    
+
+def deletesupportstaffpayment(request):
+    if request.method == 'POST':
+        paymentid = request.POST.get('paymentid')
+        payid = Supportstaffpayment.objects.filter(paymentid = paymentid)
+        payid.delete()
+        messages.success(request , "Payment deleted successfully")
+        supportstaffinfo = Supportstaffpayment.objects.all()
+        return redirect('SupportstaffpaymentsLists')
+
 def financeaddStaffpayments(request):
     if request.method == 'POST':
         supportstaffid = request.POST.get('support-staffid')
@@ -89,20 +84,18 @@ def financeaddStaffpayments(request):
         paymentdate = request.POST.get('datepaid')
         salary = request.POST.get('salary')
         amount = request.POST.get('amount')
-        
-        
+
         Supportstaffpayment.objects.create(
             supportstaffid = supportstaffid ,
             salary = salary ,
             amountpaid = amount ,
-            balance = balance ,
+            #balance = balance ,
             supportstaffname = supportstaffname,
-            paymentdate = datepaid,
-            
+            #paymentdate = datepaid,
         )
         Supportstaffpayment.save()
     return render(request , 'finance/staffpayments/financesupportstaffpaymentsList.html')
-        
+
 def receipts(request):
     if request.method == 'POST':
         receiptnum = request.POST.get('receiptnum')
@@ -111,7 +104,7 @@ def receipts(request):
         item = request.POST.get('item')
         balance = request.POST.get('balance')
         payername = request.POST.get('payername')
-        
+
         Receipts.objects.create(
             receiptnum = receiptnum ,
             transactiondate = transactiondate ,
@@ -120,7 +113,7 @@ def receipts(request):
             balance = balance ,
             payername = payername ,
         )
-        
+
         Receipts.save()
     return render(request, 'finace/financedashboard.html')
 
@@ -130,14 +123,14 @@ def bankdetails(request):
         bankname = request.POST.get('bankname')
         accnum = request.POST.get('accnum')
         accname = request.POST.get('accname')
-        
+
         Bankdetails.objects.create(
             staffname = staffname ,
             bankname = bankname ,
             accnum = accnum ,
             accname = accname
         )
-        
+
         Bankdetails.save()
     return render(request , 'finance/financedashboard.html')
 
@@ -185,7 +178,6 @@ def expenserecords(request):
         ExpenseRecord.save()
     return render(request , 'finance/financedashboard.html')
 
-
 # Create your views here.
 def financedashboard(request):
     expenses = ExpenseRecord.objects.all()
@@ -200,9 +192,6 @@ def financedashboard(request):
 
     }
     return render(request, "finance/financedashboard.html", context)
-
-    
-
 
 # fees views
 def financeaddFees(request):
@@ -239,7 +228,6 @@ def financeaddFees(request):
     students = Student.objects.all()
     fees_structures = Feesstructure.objects.all()
     return render(request, 'finance/fees/financeaddFees.html', {'students': students, 'fees_structures': fees_structures})
-
 
 def financefeesList(request):
     total_amount = Fees.objects.aggregate(Sum('amount'))['amount__sum']
@@ -278,8 +266,6 @@ def edit_std_fees(request):
         return redirect('Fees List')  # Adjust this to the correct URL name
 # fees views
 
-# feesstructure views
-
 def financeaddFeesstructure(request):
     if request.method == 'POST':
         classname = request.POST.get('classname')
@@ -311,7 +297,6 @@ def deletefeesstructure(request, feesstructureid):
         messages.success(request, 'Fees Structure deleted successfully.')
     except Feesstructure.DoesNotExist:
         messages.error(request, 'Fees Structure not found.')
-    
     return redirect('Fees Structure List')
 
 def editfeesstructure(request, feesstructureid):
@@ -320,21 +305,16 @@ def editfeesstructure(request, feesstructureid):
         if request.method == 'POST':
             updated_classname = request.POST.get('classname')
             updated_amount = request.POST.get('amount')
-            
             fees_structure.classname = updated_classname
             fees_structure.amount = updated_amount
             fees_structure.save()
             messages.success(request, 'Fees Structure updated successfully.')
             return redirect('Fees Structure List')
-            
         context = {'fees_structure': fees_structure}
         return render(request, 'finance/feesstructure/editfeesstructure.html', context)
     except Feesstructure.DoesNotExist:
         messages.error(request, 'Fees Structure not found.')
         return redirect('Fees Structure List')
-
-
-# feesstructure views
 
 # teacherpayments views
 def financeaddTeacherpayments(request):
@@ -388,7 +368,7 @@ def financeaddExpenses(request):
         expensedate = request.POST.get('expensedate')
         amountpaid = request.POST.get('amountpaid')
         balance = request.POST.get('balance')
-        
+
         expense_record = ExpenseRecord(
             category=category,
             amountrequired=amountrequired,
@@ -399,7 +379,7 @@ def financeaddExpenses(request):
         expense_record.save()
         messages.success(request, 'Expense added successfully.')  # Display a success message
         return redirect('Add Expenses')  # Redirect to expenses list page
-    
+
     return render(request, 'finance/expenses/financeaddExpenses.html')
 
 def financeexpensesList(request):
@@ -424,14 +404,13 @@ def delete_expense(request, expenseid):
 def edit_expense(request, expenseid):
     try:
         expense = ExpenseRecord.objects.get(expenseid=expenseid)
-        
+
         if request.method == 'POST':
             updated_category = request.POST.get('category')
             updated_amountrequired = request.POST.get('amountrequired')
             updated_expensedate = request.POST.get('expensedate')
             updated_amountpaid = request.POST.get('amountpaid')
             updated_balance = request.POST.get('balance')
-            
             expense.category = updated_category
             expense.amountrequired = updated_amountrequired
             expense.expensedate = updated_expensedate
@@ -440,10 +419,10 @@ def edit_expense(request, expenseid):
             expense.save()
             messages.success(request, 'Expense updated successfully.')
             return redirect('Add Expenses')
-        
+
         context = {'expense': expense}
         return render(request, 'finance/expenses/edit_expense.html', context)
-    
+
     except ExpenseRecord.DoesNotExist:
         messages.error(request, 'Expense not found.')
         return redirect('Expenses List')
@@ -455,10 +434,6 @@ def financeReports(request):
 def financeStatistics(request):
     return render(request, 'finance/financeStatistics.html')
 
-
-from django.core import serializers
-from django.http import JsonResponse
-
 def get_stdclass(request, stdnumber):
     try:
         student = Student.objects.get(stdnumber=stdnumber)
@@ -466,7 +441,7 @@ def get_stdclass(request, stdnumber):
 
         fees_structure = Feesstructure.objects.get(classname = classname)
         amount = fees_structure.amount
-        
+
         class_data = {
             'classname': classname,
             'amount': amount
@@ -475,17 +450,29 @@ def get_stdclass(request, stdnumber):
     except Student.DoesNotExist:
         return JsonResponse({}, status=404)
 
-
-
 def get_staff_salary(request,id):
     support_staff = Supportstaff.objects.get(pk=id)
-
     salary = support_staff.salary
-
     staff_data={
         'salary':salary,
     }
     return JsonResponse(staff_data)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
